HBASE-17611 Thrift2 per-call latency metrics are capped at ~2 seconds

+++ hbase-thrift/src/main/java/org/apache/hadoop/hbase/thrift2/ThriftHBaseServiceHandler.java
@@ -121,7 +121,7 @@ public class ThriftHBaseServiceHandler implements THBaseService.Iface {
         metrics.exception(e);
         throw new RuntimeException("unexpected invocation exception: " + e.getMessage());
       } finally {
-        int processTime = (int) (now() - start);
+        long processTime = now() - start;
         metrics.incMethodTime(m.getName(), processTime);
       }
       return result;
